refactor package location
